requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
jinja2==3.1.5
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.9
beautifulsoup4==4.12.3
pytest==8.3.3
